Prevent erroring when reply doesn't match
This makes it possible for the system to not fail when a reply that
ISN’T in the topic system is encountered.  This could be the case if
the reply was a messages sent to the user but NOT based on an input, or
when the topic system is refreshed.